requirements.txt
# Web Framework
flask==2.3.0
flask-cors==4.0.0
flask-socketio==5.3.0
python-dotenv==1.0.0

# Database
pymongo==4.3.3
redis==4.5.4
sqlalchemy==2.0.0

# Blockchain
web3==6.0.0
eth-account==0.8.0
py-solc-x==2.0.0

# IoT
paho-mqtt==1.6.1
aiocoap==0.4.3

# Security
pyjwt==2.6.0
cryptography==39.0.0
passlib==1.7.4
bcrypt==4.0.1

# SIEM/ELK
elasticsearch==8.7.0

# Testing & Utils
requests==2.28.2
beautifulsoup4==4.11.2
selenium==4.8.0
paramiko==3.0.0
scapy==2.5.0

# Reporting
fpdf==1.7.2
jinja2==3.1.2
markdown==3.4.1
weasyprint==57.2

# Monitoring
prometheus-client==0.16.0
psutil==5.9.4
